{"types":"C","runs":138}

+++ GraphicsContest.java
@@ -22,9 +22,9 @@ public class GraphicsContest extends GraphicsProgram {
 
 	private void setUpBall() {
 		while (true) {
-			double s = 1;
-			double x = rgen.nextDouble(0 + 2*s, getWidth() - 2*s);
-			double y = rgen.nextDouble(0 + 2*s, getHeight() - 2*s);
+			int s = 1;
+			int x = rgen.nextInt(0 + 2*s, getWidth() - 2*s);
+			int y = rgen.nextInt(0 + 2*s, getHeight() - 2*s);
 			GRect pixel = new GRect (x - s, y - s, s*2, s*2);
 			pixel.setFilled(true);
 			pixel.setColor(rgen.nextColor());
